requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
bcrypt==3.1.4             # via flask-bcrypt
blinker==1.4              # via flask-mail
cffi==1.11.5              # via bcrypt, cryptography
click==6.7
cryptography==2.3         # via jwt
flask-bcrypt==0.7.1
flask-login==0.4.1
flask-mail==0.9.1
flask-marshmallow==0.9.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
idna==2.7                 # via cryptography
itsdangerous==0.24
jinja2==2.10
jwt==0.5.4
markupsafe==1.0
marshmallow-sqlalchemy==0.14.0
marshmallow==2.15.4       # via flask-marshmallow, marshmallow-sqlalchemy
pillow==5.2.0
pycparser==2.18           # via cffi
six==1.11.0               # via bcrypt, cryptography, flask-marshmallow
sqlalchemy==1.2.10        # via flask-sqlalchemy, marshmallow-sqlalchemy
werkzeug==0.14.1
wtforms==2.2.1            # via flask-wtf
